introduce time attribute

+++ setup/acc2/acc2.py
@@ -83,7 +83,7 @@ class ACC2(Veros):
     @veros_method
     def set_topography(self):
         x, y = np.meshgrid(self.xt, self.yt, indexing="ij")
-        self.kbot[...] = np.logical_or(x > 1.0, y < -20)
+        self.kbot[...] = np.logical_or(x > 1.0, y < -20).astype(np.int)
 
     @veros_method
     def set_initial_conditions(self):
@@ -119,7 +119,8 @@ class ACC2(Veros):
     def set_diagnostics(self):
         self.diagnostics["snapshot"].output_frequency = 86400 * 10
         self.diagnostics["averages"].output_variables = (
-            "salt", "temp", "u", "v", "w", "psi", "surface_taux", "surface_tauy")
+            "salt", "temp", "u", "v", "w", "psi", "surface_taux", "surface_tauy"
+        )
         self.diagnostics["averages"].output_frequency = 365 * 86400.
         self.diagnostics["averages"].sampling_frequency = self.dt_tracer * 10
         self.diagnostics["overturning"].output_frequency = 365 * 86400. / 48.
